+++ adminsortable/admin.py
@@ -10,17 +10,17 @@ from django.core.paginator import EmptyPage
 from django.db import transaction
 from django.db.models import Max, F
 from django.forms.models import BaseInlineFormSet
-from django.forms.widgets import HiddenInput
+from django.forms import widgets
 from django.http import HttpResponse, HttpResponseBadRequest, HttpResponseNotAllowed, HttpResponseForbidden
-from django.views.decorators.csrf import csrf_exempt
 from django.core.serializers.json import DjangoJSONEncoder
 from django.contrib import admin
 
 
 class SortableAdminBase(object):
-    class Media:
+    @property
+    def media(self):
         css = {'all': ('adminsortable/css/sortable.css',)}
-        if VERSION[0] == 1 and VERSION[1] <= 5:
+        if VERSION[:2] <= (1, 5):
             js = (
                 'adminsortable/js/plugins/admincompat.js',
                 'adminsortable/js/libs/jquery.ui.core-1.7.1.js',
@@ -49,6 +49,7 @@ class SortableAdminBase(object):
             except ImportError:
                 # other packages may pollute the import search path with 'cms'
                 pass
+        return super(SortableAdminBase, self).media + widgets.Media(css=css, js=js)
 
 
 class SortableAdminMixin(SortableAdminBase):
@@ -56,16 +57,19 @@ class SortableAdminMixin(SortableAdminBase):
 
     def __init__(self, model, admin_site):
         try:
-            self.default_order_field = model._meta.ordering[0].lstrip('-')
+            if model._meta.ordering[0].startswith('-'):
+                self.default_order_directions = ((1, 0), (0, 1))
+                self.default_order_field = model._meta.ordering[0].lstrip('-')
+            else:
+                self.default_order_directions = ((0, 1), (1, 0))
+                self.default_order_field = model._meta.ordering[0]
         except (AttributeError, IndexError):
             raise ImproperlyConfigured('Model {0}.{1} requires a list or tuple "ordering" in its Meta class'.format(model.__module__, model.__name__))
         super(SortableAdminMixin, self).__init__(model, admin_site)
-        if not isinstance(getattr(self, 'exclude', None), (list, tuple)):
+        if not isinstance(self.exclude, (list, tuple)):
             self.exclude = [self.default_order_field]
         elif not self.exclude or self.default_order_field != self.exclude[0]:
             self.exclude = [self.default_order_field] + self.exclude
-        if not getattr(self, 'change_list_template', None):
-            self.change_list_template = 'adminsortable/change_list.html'
         if not self.list_display_links:
             self.list_display_links = self.list_display[0]
         self._add_reorder_method()
@@ -106,12 +110,14 @@ class SortableAdminMixin(SortableAdminBase):
             self.order_by = '-' + self.default_order_field
         else:
             self.enable_sorting = False
-        if self.enable_sorting:
-            try:
-                self.Media.js += ('adminsortable/js/list-sortable.js',)
-            except AttributeError:
-                self.Media.js = ('adminsortable/js/list-sortable.js',)
         return super(SortableAdminMixin, self).get_changelist(request, **kwargs)
+
+    @property
+    def media(self):
+        m = super(SortableAdminMixin, self).media
+        if self.enable_sorting:
+            m = m + widgets.Media(js=('adminsortable/js/list-sortable.js',))
+        return m
 
     def _add_reorder_method(self):
         """
@@ -127,11 +133,18 @@ class SortableAdminMixin(SortableAdminBase):
             return html
 
         setattr(func, 'allow_tags', True)
-        setattr(func, 'short_description', _('Sort'))
+        # if the field used for ordering has a verbose name use it, otherwise default to "Sort"
+        for order_field in self.model._meta.fields:
+            if order_field.name == self.default_order_field:
+                short_description = getattr(order_field, 'verbose_name', None)
+                if short_description:
+                    setattr(func, 'short_description', short_description)
+                    break
+        else:
+            setattr(func, 'short_description', _('Sort'))
         setattr(func, 'admin_order_field', self.default_order_field)
         setattr(self, '_reorder', MethodType(func, self))
 
-    @csrf_exempt
     def update(self, request):
         if not request.is_ajax() or request.method != 'POST':
             return HttpResponseBadRequest('Not an XMLHttpRequest')
@@ -170,28 +183,32 @@ class SortableAdminMixin(SortableAdminBase):
 
     def _move_item(self, request, startorder, endorder):
         if self._get_order_direction(request) != '-1':
-            order_up, order_down = 0, 1
+            order_up, order_down = self.default_order_directions[0]
         else:
-            order_up, order_down = 1, 0
+            order_up, order_down = self.default_order_directions[1]
         if startorder < endorder - order_up:
             finalorder = endorder - order_up
             move_filter = {
-                '%s__gte' % self.default_order_field: startorder,
-                '%s__lte' % self.default_order_field: finalorder,
+                '{0}__gte'.format(self.default_order_field): startorder,
+                '{0}__lte'.format(self.default_order_field): finalorder,
             }
             order_by = self.default_order_field
             move_update = {self.default_order_field: F(self.default_order_field) - 1}
         elif startorder > endorder + order_down:
             finalorder = endorder + order_down
             move_filter = {
-                '%s__gte' % self.default_order_field: finalorder,
-                '%s__lte' % self.default_order_field: startorder,
+                '{0}__gte'.format(self.default_order_field): finalorder,
+                '{0}__lte'.format(self.default_order_field): startorder,
             }
             order_by = '-' + self.default_order_field
             move_update = {self.default_order_field: F(self.default_order_field) + 1}
         else:
             return self.model.objects.none()
-        with transaction.commit_on_success():
+        if VERSION[:2] <= (1, 5):
+            atomic_context_manager = transaction.commit_on_success
+        else:
+            atomic_context_manager = transaction.atomic
+        with atomic_context_manager():
             obj = self.model.objects.get(**{self.default_order_field: startorder})
             setattr(obj, self.default_order_field, self.get_max_order() + 1)
             obj.save()
@@ -249,7 +266,7 @@ class CustomInlineFormSet(BaseInlineFormSet):
             raise ImproperlyConfigured('Model {0}.{1} requires a list or tuple "ordering" in its Meta class'.format(self.model.__module__, self.model.__name__))
         self.form.base_fields[self.default_order_field].is_hidden = True
         self.form.base_fields[self.default_order_field].required = False
-        self.form.base_fields[self.default_order_field].widget = HiddenInput()
+        self.form.base_fields[self.default_order_field].widget = widgets.HiddenInput()
         super(CustomInlineFormSet, self).__init__(*args, **kwargs)
 
     def save_new(self, form, commit=True):
@@ -273,15 +290,15 @@ class CustomInlineFormSet(BaseInlineFormSet):
 class SortableInlineAdminMixin(SortableAdminBase):
     formset = CustomInlineFormSet
 
-    def __init__(self, parent_model, admin_site):
-        version = (VERSION[0] == 1 and VERSION[1] <= 5) and '1.5' or '1.6'
+    @property
+    def media(self):
+        return super(SortableInlineAdminMixin, self).media + widgets.Media(js=('adminsortable/js/inline-sortable.js',))
+
+    @property
+    def template(self):
+        version = VERSION[:2] <= (1, 5) and '1.5' or '1.6'
         if isinstance(self, admin.StackedInline):
-            self.template = 'adminsortable/stacked-%s.html' % version
-            self.Media.js += ('adminsortable/js/inline-sortable.js',)
-        elif isinstance(self, admin.TabularInline):
-            self.template = 'adminsortable/tabular-%s.html' % version
-            self.Media.js += ('adminsortable/js/inline-sortable.js',)
-        else:
-            raise ImproperlyConfigured('Class {0}.{1} must also derive from admin.TabularInline or admin.StackedInline'.format(self.__module__, self.__class__))
-        self.Media.css['all'] += ('adminsortable/css/sortable.css',)
-        super(SortableInlineAdminMixin, self).__init__(parent_model, admin_site)
+            return 'adminsortable/stacked-{0}.html'.format(version)
+        if isinstance(self, admin.TabularInline):
+            return 'adminsortable/tabular-{0}.html'.format(version)
+        raise ImproperlyConfigured('Class {0}.{1} must also derive from admin.TabularInline or admin.StackedInline'.format(self.__module__, self.__class__))