+++ adminsortable/static/adminsortable/js/list-sortable.js
@@ -18,8 +18,10 @@ jQuery.extend({
 
 // make list view sortable
 jQuery(function($) {
-	var sortable_update_url = $('#adminsortable_update_url').attr('href');
+	var sortable_update_url = $('#adminsortable_update_url').attr('href') || 'adminsortable_update/';
 	var startorder, endorder;
+	var csrfvalue = $('#changelist-form').find('input[name="csrfmiddlewaretoken"]').val();
+
 	$('#result_list').sortable({
 		handle: 'div.drag',
 		items: 'tr',
@@ -51,7 +53,8 @@ jQuery(function($) {
 				data: {
 					o: $.getQueryParam('o'),
 					startorder: startorder,
-					endorder: endorder
+					endorder: endorder,
+					csrfmiddlewaretoken: csrfvalue
 				},
 				success: function(moved_items) {
 					$.each(moved_items, function(index, item) {